% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_BOD.R
\name{conditional_BOD}
\alias{conditional_BOD}
\title{Conditional BOD function}
\usage{
conditional_BOD(
  output,
  exogenous = FALSE,
  m,
  B,
  RTS,
  ORIENTATION,
  similarity = FALSE,
  inclusion = FALSE,
  print = FALSE
)
}
\arguments{
\item{output}{matrix (or vector) of indicators along which the units are evaluated.}

\item{exogenous}{matrix (or vector) of exogenous variables involved in the conditional analysis. The similarity among the units is determined according to the exogeneous variable(s) using the function npudensbw and npudens (from the package np) with epanechnikov kernel.}

\item{m}{number of unit to be included in the reference set}

\item{B}{number of bootstrap replicates}

\item{RTS}{For more details see the dea function in the package Benchmarking. Text string or a number defining the underlying DEA technology / returns to scale assumption.
0    fdh    Free disposability hull, no convexity assumption
1    vrs    Variable returns to scale, convexity and free disposability
2    drs    Decreasing returns to scale, convexity, down-scaling and free disposability
3    crs    Constant returns to scale, convexity and free disposability
4    irs    Increasing returns to scale, (up-scaling, but not down-scaling), convexity and free disposability
5    irs2    Increasing returns to scale (up-scaling, but not down-scaling), additivity, and free disposability
6    add    Additivity (scaling up and down, but only with integers), and free disposability; also known af replicability and free disposability, the free disposability and replicability hull (frh) -- no convexity assumption
7    fdh+    A combination of free disposability and restricted or local constant return to scale
10    vrs+    As vrs, but with restrictions on the individual lambdas via param}

\item{ORIENTATION}{For more details see the dea function in the package Benchmarking. Input efficiency "in" (1), output efficiency "out" (2), and graph efficiency "graph" (3). For use with DIRECT, an additional option is "in-out" (0).}

\item{similarity}{matrix of similarities. In alternative to provide the exogenous variables, the matrix of similarities can be directly provided. This allow to customize the estimation of the similarities.}

\item{inclusion}{If inclusion = TRUE the unit under analysis is included in the reference set. So, no super efficient scores are allowed. By default inclusion = FALSE.}

\item{print}{If print = TRUE the number of the unit under evaluation is printed. In case of large sample the function could require some time, so it could be useful to control how many units have already been evaluated and which one still have to be evaluated. By default print = FALSE.}
}
\description{
This function allows to compute Robust and Conditional BOD scores.
}
\examples{
conditional_BOD()

}
\keyword{BOD,}
\keyword{Conditional}
