% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separability_test.R
\name{separability_test}
\alias{separability_test}
\title{Separability condition test}
\usage{
separability_test(rDEA, cDEA, p, q, B, model, alpha = 0.05, print = TRUE)
}
\arguments{
\item{rDEA}{vector containing the robust DEA (or BOD or FDH) scores}

\item{cDEA}{vector containing the conditional DEA (or BOD or FDH) scores}

\item{p}{number input used in the analysys (in case of BOD analysis set p=1)}

\item{q}{number output used in the analysis}

\item{B}{number of bootstrap replicates}

\item{model}{model implemented to obtain the efficiency score. Possible values for this parameter are "DEA" or "FDH"}

\item{alpha}{significance level. By default alpha = 0.05}

\item{print}{if print = TRUE the function print the result of the test. By default print = TRUE.}
}
\description{
This function test the separability condition and return the value of t and the p-value.  The main reference is the paper by Daraio and Simar (2015)
}
\examples{
separability_test()

}
\keyword{condition,}
\keyword{separability}
\keyword{test,}
